Add Microsoft.CodeAnalysis.Extensions package
