+++ eng/config/PublishData.json
@@ -36,6 +36,7 @@
       "Microsoft.CodeAnalysis.PooledObjects": "arcade",
       "Microsoft.CodeAnalysis.Collections": "arcade",
       "Microsoft.CodeAnalysis.Threading": "arcade",
+      "Microsoft.CodeAnalysis.Extensions": "vs-impl",
       "Microsoft.CodeAnalysis.Features": "arcade",
       "Microsoft.CodeAnalysis.EditorFeatures": "vssdk",
       "Microsoft.CodeAnalysis.EditorFeatures.Common": "vssdk",
