+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/AccessibilityUtilities.cs
