+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/IAssemblySymbolExtensions.cs
