+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/INamespaceOrTypeSymbolExtensions.cs
