+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/IPropertySymbolExtensions.cs
