+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/ISymbolExtensions.RequiresUnsafeModifierVisitor.cs
