+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/ITypeGenerator.cs
