+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/ITypeParameterSymbolExtensions.cs
