+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/ITypeSymbolExtensions.CollectTypeParameterSymbolsVisitor.cs
