+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/MethodKindExtensions.cs
