+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/PredefinedOperator.cs
