+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/PredefinedType.cs
