+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/SignatureComparer.cs
