+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/SymbolEquivalenceComparer.SignatureTypeSymbolEquivalenceComparer.cs
