+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Extensions/Symbols/SymbolVisibility.cs
